Apply suggestions from code review

+++ src/types.hpp
@@ -8,7 +8,7 @@ namespace gabe {
 using uint8 = unsigned char;
 using uint16 = unsigned short int;
 using uint32 = unsigned int;
-using uint64 = unsigned long long int;
+using uint64 = unsigned long int;
 
 using Size = uint64;
 } // namespace gabe
